test: move discover tests
